refactor: 컨벤션에 맞는 형태로 파일 수정
