+++ src/components/Badge/Badge.stories.tsx
@@ -1,7 +1,7 @@
+import { BADGE_TYPE } from '@constants/badge';
 import type { Meta, StoryObj } from '@storybook/react';
 
-import Badge from './Badge';
-import { BADGE_TYPE } from './constants';
+import Badge from '.';
 
 const meta: Meta<typeof Badge> = {
   title: 'Components/Badge',
