+++ src/components/Modal/Modal.stories.tsx
@@ -1,4 +1,4 @@
-import { useModalStore } from '@stores/modalStore';
+import { useModalStore } from '@stores/modal';
 import type { Meta, StoryObj } from '@storybook/react';
 
 import Modal from '.';
