+++ src/components/Modal/style.css.ts
@@ -1,7 +1,8 @@
+import { createVar, style } from '@vanilla-extract/css';
+
 import { sprinkles } from '@styles/sprinkles.css';
 import { COLORS } from '@styles/tokens';
 import { middleLayer, positionCenter, topLayer } from '@styles/utils.css';
-import { createVar, style } from '@vanilla-extract/css';
 
 export const modalStyle = style([
   positionCenter,
