+++ src/components/Tabs/components/TabsList.tsx
@@ -1,6 +1,6 @@
 import type { PropsWithChildren } from 'react';
 
-import * as styles from './style.css';
+import * as styles from '../style.css';
 
 const TabsList = ({ children }: PropsWithChildren) => {
   return <ul className={styles.tabsList}>{children}</ul>;
