+++ src/components/Tabs/components/TabsTrigger.tsx
@@ -1,7 +1,7 @@
 import type { PropsWithChildren } from 'react';
 
-import { useTabsContext } from '../../hooks/useTabsContext';
-import * as styles from './style.css';
+import { useTabsContext } from '../../../hooks/useTabsContext';
+import * as styles from '../style.css';
 
 interface TabsTriggerProps {
   value: string;
