+++ src/components/Tabs/style.css.ts
@@ -1,7 +1,8 @@
-import { COLORS } from '@styles/tokens';
-import * as utils from '@styles/utils.css';
 import { style } from '@vanilla-extract/css';
 import { recipe } from '@vanilla-extract/recipes';
+
+import { COLORS } from '@styles/tokens';
+import * as utils from '@styles/utils.css';
 
 export const tabsList = style([
   utils.flexCenter,
