+++ src/constants/toast.ts
@@ -1,0 +1,4 @@
+export const TOAST_DURATION_TIME = {
+  SHOW: 3500,
+  ACTION: 5000,
+} as const;
