+++ src/models/toast.ts
@@ -1,7 +1,4 @@
-export const TOAST_DURATION_TIME = {
-  SHOW: 3500,
-  ACTION: 5000,
-} as const;
+import type { TOAST_DURATION_TIME } from '@constants/toast';
 
 export type ToastDurationTime =
   (typeof TOAST_DURATION_TIME)[keyof typeof TOAST_DURATION_TIME];
